[add] enemy

+++ TSGame/Shape.h
@@ -4,7 +4,6 @@
 #include <cmath>
 #include "Vector2.h"
 
-//基本的にテクスチャ乗っけるまでのデバグ用なのれす
 namespace Shape {
 	struct ShapeVertex {
 		D3DXVECTOR3 p;
@@ -22,7 +21,7 @@ namespace Shape {
 		line->Draw(vec, 2, color);
 		line->End();
 	}
-	inline void drawCircle(LPDIRECT3DDEVICE9 d3dDevice, Vector2 pos, int radius, D3DXCOLOR color = 0xFFFFFFFF) {
+	inline void drawCircle(LPDIRECT3DDEVICE9 d3dDevice, Vector2 pos, float radius, D3DXCOLOR color = 0xFFFFFFFF) {
 		static const int splitNum = 64;
 		std::vector<ShapeVertex> vtx(splitNum, { { 0, 0, 0 }, 1, color });
 		for (int i = 0; i < splitNum; i++) {
@@ -48,7 +47,7 @@ namespace Shape {
 		d3dDevice->SetFVF(D3DFVF_XYZRHW | D3DFVF_DIFFUSE);
 		d3dDevice->DrawPrimitiveUP(D3DPT_TRIANGLESTRIP, 2, vtx.data(), sizeof(ShapeVertex));
 	}
-	inline void drawNgon(LPDIRECT3DDEVICE9 d3dDevice, Vector2 pos, int num, int radius, float radian, D3DXCOLOR color = 0xFFFFFFFF) {
+	inline void drawNgon(LPDIRECT3DDEVICE9 d3dDevice, Vector2 pos, int num, float radius, float radian, D3DXCOLOR color = 0xFFFFFFFF) {
 		std::vector<ShapeVertex> vtx(num, { { 0, 0, 0 }, 1, color });
 		for (int i = 0; i < num; i++) {
 			float tRadian = D3DX_PI * 2 * i / num;
